diploma_backend/requirements.txt
Django==6.0.1
djangorestframework==3.16.1
django-cleanup==9.0.0
django-filter==25.2
django-phonenumber-field==8.4.0
phonenumbers==9.0.24
django-credit-cards==0.4.1
drf-spectacular==0.29.0
drf-yasg==1.21.14
Pillow==12.1.0
python-dotenv==1.2.1
gunicorn==23.0.0
sentry-sdk==2.50.0
# Testing
pytest==9.0.2
pytest-django==4.11.1
pytest-asyncio==0.26.0